deal with not responding nameservers (offline or unreachable)
if a nameserver does not respond, we flag it in the database
(available = False) and stop talking to it for 5 minutes.

after 5 minutes, we flag it available = True again and retry talking to it
(and if not successful, flag it as unavailable again).

this avoids a unresponsive UI as we do a lot of ns queries from the templates
to get v4 and v6 addresses of our hosts from master nameservers.

we log a warning when setting unavailable and a info when setting available.