remove help view, move project info to docs, make about view use settings
help: there was only a link to the rtd documentation (now directly linked from a "Documentation" navbar entry)
and some duplicate router configuration that was also shown (even with the correct values) when adding a host.

the help page html was not valid due to the values inserted into the router configuration help.

help/documentation policy: add static help to the docs, add dynamic help directly at the places (in the views) where needed