partially revert commit 78b27f65c0a90eb51fba09112e9bc681b8070b1a
the host.get_ipv4 (and same for get_ipv6) call was duplicated, resulting in
additional ns queries.

also, the semantics of this part of the change was unclear or even erroneous.
get_ipv4 returns either the ip address or 'error', so that "if" would always
gets a trueish value to evaluate, making it pointless.