do not set A or AAAA record if IFID field is blank
in fact, even delete the A or AAAA record if the field is blank and an update is received.

sometimes it makes sense to have only ipv6 records in DNS if you only have 1 public ipv4.
then you can reach your devices without getting (maybe wrong or not port-forwarded) v4 addresses on lookup.