rewrite socket error handling, fixes #522
socket.error is deprecated and is the same as OSError since python 3.3.
Of course OSError can mean a lot of stuff, so not sure how that goes...

socket.herror is for gethostbyaddr and has some documented h_error values in netdb.h
(but, AFAIK, these symbols are not available via python).

added a handler for h_error 0 (#522), which crashed since a while.
removed some unclear error values, let's see what happens.